Add InitFinalize base class of MPI_InitFinalize to initialize/finalize deal.II and other exteranl properties
